test: move focus recovery coverage to swift testing
